+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is roughly based on [Keep a Changelog](https://keepachangelog.com/en/
 ## Work in progress: 0.46.0
 
 - Add `AggregatorBackendConfig.processes_target_version` to define processes target version ([#149](https://github.com/Open-EO/openeo-aggregator/issues/149))
+- Fixup compatibility with `openeo_driver` 0.132.0 ([Open-EO/openeo-python-driver#382](https://github.com/Open-EO/openeo-python-driver/issues/382))
 
 
 ## 0.45.0