+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is roughly based on [Keep a Changelog](https://keepachangelog.com/en/
 
 <!-- start-of-changelog -->
 
-## Work in progress: 0.46.0
+## Work in progress: 0.47.0
+
+
+## 0.46.0
 
 - Add `AggregatorBackendConfig.processes_target_version` to define processes target version ([#149](https://github.com/Open-EO/openeo-aggregator/issues/149))
 - Fixup compatibility with `openeo_driver` 0.132.0 ([Open-EO/openeo-python-driver#382](https://github.com/Open-EO/openeo-python-driver/issues/382))